+++ lib/amazon_business_api/resources/place_order_attribute.rb
@@ -1,10 +1,10 @@
 # frozen_string_literal: true
 
-require_relative 'attributes/product_reference'
-require_relative 'attributes/buying_option_reference'
+require_relative 'place_order_attributes/product_reference'
+require_relative 'place_order_attributes/buying_option_reference'
 
 module AmazonBusinessApi
-  class Attribute < AmazonBusinessApi::Resource
+  class PlaceOrderAttribute < AmazonBusinessApi::Resource
     # https://developer-docs.amazon.com/amazon-business/docs/reporting-api-v1-reference-1
     attribute :attribute_type, type: LedgerSync::Type::String
     references_one :product_reference, to: ProductReference
