+++ lib/amazon_business_api/resources/place_order_attributes/buying_option_reference.rb
