+++ lib/amazon_business_api/resources/place_order_attributes/product_reference.rb
