+++ lib/amazon_business_api/resources/place_order_expectation.rb
@@ -1,9 +1,9 @@
 # frozen_string_literal: true
 
-require_relative 'expectations/amount'
+require_relative 'place_order_expectations/amount'
 
 module AmazonBusinessApi
-  class Expectation < AmazonBusinessApi::Resource
+  class PlaceOrderExpectation < AmazonBusinessApi::Resource
     # https://developer-docs.amazon.com/amazon-business/docs/reporting-api-v1-reference-1
     attribute :expectation_type, type: LedgerSync::Type::String
     references_one :amount, to: Amount
