+++ lib/amazon_business_api/resources/place_order_expectations/amount.rb
