+++ lib/amazon_business_api/resources/request_line_item.rb
@@ -1,13 +1,13 @@
 # frozen_string_literal: true
 
-require_relative 'attribute'
-require_relative 'expectation'
+require_relative 'place_order_attribute'
+require_relative 'place_order_expectation'
 
 module AmazonBusinessApi
   class RequestLineItem < AmazonBusinessApi::Resource
     # https://developer-docs.amazon.com/amazon-business/docs/reporting-api-v1-reference-1
     attribute :quantity, type: LedgerSync::Type::Integer
-    references_many :attributes, to: Attribute
-    references_many :expectations, to: Expectation
+    references_many :place_order_attributes, to: PlaceOrderAttribute
+    references_many :place_order_expectations, to: PlaceOrderExpectation
   end
 end
